+++ .gitignore
@@ -1,2 +1,3 @@
 TODO
 timeline.md
+help.log
